Move the LocalTime implementation to gtest-port.h
Move the LocalTime implementation to gtest-port.h so it can be customized for each port
